+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## [v1.1.2--snapshot]
+## [v1.2.1--snapshot]
+
+### Added
+- New invoker mechanism
+
+### Changed
+- 
+
+### Fixed
+- 
+
+## [v1.2.0]
 
 ### Added
 - New local_executor() to run pywren jobs in the local machine
@@ -8,6 +19,7 @@
 - New localhost storage backend
 - New docker_executor() to run pywren jobs in the local machine by using docker
 - New docker compute backend
+
 
 ### Changed
 - Docs updated